// src/pages/Projects/DccTranslation/DccTranslation.types.ts
export type ContentBlock =
  | {
      type: "paragraph";
      text: string;
    }
  | {
      type: "ordered-list";
      items: ContentListItem[];
    }
  | {
      type: "unordered-list";
      items: ContentListItem[];
    };

export type ContentListItem = {
  title?: string;
  text: string;
};

export type ProjectContentSection = {
  title: string;
  blocks: ContentBlock[];
};